Change the way random graphs are generated
This commit adds a new command line argument (-N) that
specifies the maximum number of edges that should be connected
to each node of a random graph. The actual number of edges
is a uniform distribution between 0 and the given value.

This makes random graph generation much faster in most
circumstances and provides a graph that has more variety in
terms of number of edges connected to a node